+++ usr.sbin/amd/amd/util.c
@@ -36,7 +36,7 @@
  * SUCH DAMAGE.
  *
  *	from: @(#)util.c	8.1 (Berkeley) 6/6/93
- *	$Id: util.c,v 1.2 1996/08/10 21:39:43 deraadt Exp $
+ *	$Id: util.c,v 1.3 1997/01/31 14:42:02 graichen Exp $
  */
 
 /*
@@ -242,10 +242,10 @@ char **chp;
  * addr is in network byte order.
  * sizeof(buf) needs to be at least 16.
  */
-char *inet_dquad P((char *buf, unsigned long addr));
+char *inet_dquad P((char *buf, u_int32_t addr));
 char *inet_dquad(buf, addr)
 char *buf;
-unsigned long addr;
+u_int32_t addr;
 {
 	addr = ntohl(addr);
 	sprintf(buf, "%d.%d.%d.%d",
